Merged PR 2009: Setup VS Code Devcontainers for both Frontend and Backend.
The VS Code extension adds a folder called .devcontainers which is hidden by default.  One was added in the frontend root folder and another in the backend root folder.

I choose not to add documentation for the following reasons:
- I believe adding more documentation on yet another way to run the app will be confusing as attendee feedback stated that there was a lot of docs
- Devcontainers has not been properly vetted
- Due to multiple OSs, it is impossible to completely automate getting the Integration Tests to work inside the container.  There are manual steps that have to be done at this stage.

As stated, VS Code Devcontainers created net new folders, so I think it is okay to keep what is created but if we want to abandon this PR, I am fine with it.  Ultimately, I don't think it was a waste because some good sharing opportunities & feedback will result from this PR research.

**Instructions**
- Install vscode
- Install [devcontainers](https://marketplace.visualstudio.com/items?itemName=ms-vscode-remote.vscode-remote-extensionpack) extension
- Create .env.development.local in urlist/frontend
- Create appsetttings.Development.json in urlist/api/src/LinkLink
- Open up a blank vscode environment, hit devcontainers icon at bottom left, choose “open folder in container”, open in urlist/frontend
  - Once finished loading, npm install
  - npm run serve
- File, new window, hit icon at bottom left, open urlist/api/src/LinkLink
  - Once finished loading, dotnet run
- Frontend=http://localhost:8080
- Backend=http://localhost:5000

Hit the icon on bottom left again and choose Close Connection to shutdown

**Note:** As explained in the 2/28 Demo, lots of manual steps to get the Integration Tests to work inside the container due to [dotnet dev-certs limitations](https://github.com/dotnet/aspnetcore/issues/7246).